+++ managers/oclcCatalog.py
@@ -2,6 +2,13 @@ import os
 import requests
 from requests.exceptions import Timeout, ConnectionError
 
+from logger import createLog
+from managers.oclcAuth import OCLCAuthManager
+
+
+logger = createLog(__name__)
+
+
 class OCLCCatalogManager:
     CATALOG_URL = 'http://www.worldcat.org/webservices/catalog/content/{}?wskey={}'
     def __init__(self):
@@ -18,13 +25,35 @@ class OCLCCatalogManager:
         try:
             classifyResp = requests.get(catalogQuery, timeout=3)
         except (Timeout, ConnectionError):
-            print('Failed to query URL {}'.format(catalogQuery))
-            return self.queryCatalog(oclcNo)
+            logger.warn('Failed to query URL {}'.format(catalogQuery))
+            return self.queryCatalog()
 
         if classifyResp.status_code != 200:
-            print('OCLC Catalog Request failed with status {}'.format(
+            logger.warn('OCLC Catalog Request failed with status {}'.format(
                 classifyResp.status_code
             ))
             return None
-    
+
         return classifyResp.text
+
+    def queryBriefBibs(self, query):
+        """Accepts a query in the form of an OCLC keyword search or fielded search"""
+        token = OCLCAuthManager.getToken()
+        bibsEndpoint = self.OCLC_SEARCH_URL + 'brief-bibs'
+
+        headers = {"Authorization": f"Bearer {token}"}
+        try:
+            bibsResp = requests.get(
+                bibsEndpoint,
+                headers=headers,
+                params={'q': query}
+            )
+        except (Timeout, ConnectionError):
+            logger.warn(f'Failed to query {bibsEndpoint} with query {query}')
+            return self.queryCatalog()
+
+        if bibsResp.status_code != 200:
+            logger.warn(f'OCLC Catalog Request failed with status {bibsResp.status_code}')
+            return None
+
+        return bibsResp.json()